Update solution for tanh-activation

+++ tanh-activation/tanh-activation.py
@@ -1,6 +1,6 @@
 import numpy as np
 
 def tanh(x):
-    a = np.asarray(x)
+    a = np.asarray(x, dtype = float)
     a = (np.exp(a) - np.exp(-a)) / (np.exp(a) + np.exp(-a))
     return a
